% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_m3u.R
\name{add_m3u}
\alias{add_m3u}
\title{Create a M3U File from a list of URL}
\usage{
add_m3u(data, podcast, path = ".", na_rm = TRUE)
}
\arguments{
\item{data}{a \code{data.frame} created by \code{get_metadata()}.}

\item{podcast}{a character of length 1. The name of the podcast.}

\item{path}{a character of length 1. The folder to save the M3U file.}

\item{na_rm}{a logical. If \code{TRUE} (default) remove episodes with incomplete
information.}
}
\value{
The content of the M3U file (optional).
}
\description{
Creates a M3U (MPEG version 3.0 URL) file from a list of URLs. This file can
be read by any multimedia reader (e.g. VLC). These audio/video URLs will be
streamed (not downloaded).
For more information: \url{https://en.wikipedia.org/wiki/M3U}
}
\examples{
\dontrun{
## Create a folder to store results ----

path <- "Podcasts"
dir.create(path)


## Retrieve episodes information ----

tab <- get_metadata("un-ete-avec-homere", path)


## Create a M3U playlist ----

x <- add_m3u(tab, "un-ete-avec-homere", path)

cat(paste0(x, collapse = "\n"))
}
}
