% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_m3u.R
\name{all_m3u}
\alias{all_m3u}
\title{Create a M3U File from a list of URL}
\usage{
all_m3u(path = ".", na_rm = TRUE)
}
\arguments{
\item{path}{the path to save the M3U file (must exist).}

\item{na_rm}{a logical. If \code{TRUE} (default) remove episodes with incomplete
information.}
}
\value{
The content of the M3U file (optional).
}
\description{
Creates a M3U (MPEG version 3.0 URL) file from a list of URLs. This file can
be read by any multimedia reader (e.g. VLC). These audio/video URLs will be
streamed (not downloaded).
For more information: \url{https://en.wikipedia.org/wiki/M3U}
}
\examples{
\dontrun{
## Create a M3U playlist ----
x <- all_m3u(path)

cat(paste0(x, collapse = "\n"))
}
}
