% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_m3u.R
\name{create_m3u}
\alias{create_m3u}
\title{Create a m3u file from a list of URL}
\usage{
create_m3u(data, podcast, path = ".")
}
\arguments{
\item{data}{a \code{data.frame} created by \code{get_metadata()}.}

\item{podcast}{a character of length 1. The name of the podcast.}

\item{path}{a character of length 1. The folder to save the M3U file.}
}
\value{
The content of the M3U file (optional).
}
\description{
Creates a m3u (MPEG version 3.0 URL) file from a list of URLs. This file can
be read by any multimedia reader (e.g. VLC). These audio/video URLs will be
streamed (not downloaded).
For more information: \url{https://en.wikipedia.org/wiki/M3U}
}
\examples{
\dontrun{
## Retrieve episodes metadata ----

tab <- get_metadata("un-ete-avec-homere", radio = "franceinter")


## Create a m3u playlist ----

x <- create_m3u(tab, "un-ete-avec-homere")
cat(paste0(x, collapse = "\n"))
}
}
\seealso{
get_metadata()
}
