% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franceinter-package.R
\docType{package}
\name{franceinter-package}
\alias{franceinter}
\alias{franceinter-package}
\title{franceinter: Get France Inter Podcasts}
\description{
Retrieves Radio France (France Inter, France Culture, etc.) podcasts metadata (date, title, mp3 url, and duration). User can easily download metadata for any podcast. The name of the podcast must be extracted from the website URL (e.g. Le Moment Meurice must be written as le-moment-meurice).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ahasverus/franceinter}
  \item Report bugs at \url{https://github.com/ahasverus/franceinter/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@gmail.com} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
