% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Retrieve Podcast Information}
\usage{
get_metadata(
  podcast,
  start_date = NULL,
  end_date = NULL,
  path = ".",
  na_rm = TRUE
)
}
\arguments{
\item{podcast}{a character of length 1. The name of the M3U file.}

\item{start_date}{a character of length 1. The date of the oldest episode
to retrieve. If \code{NULL} the \code{"2010-01-01"} will be used.}

\item{end_date}{a character of length 1. The date of the most recent episode
to retrieve. If \code{NULL} the current date will be used.}

\item{path}{the path to save metadata as CSV (must exist).}

\item{na_rm}{a logical. If \code{TRUE} (default) remove episodes with incomplete
information.}
}
\value{
A four-columns data frame with:
\itemize{
\item \code{date}: the date of the episode;
\item \code{title}: the title of the episode;
\item \code{duration}: the duration of the episode (in seconds);
\item \code{file_url}: the URL of the mp3.
}
}
\description{
Retrieves podcast information (i.e. date, title, mp3 duration and url).
Information will be saved as a CSV file. Each time this function is called
only new episodes will be treated.
}
\examples{
\dontrun{
## Create a folder to store results ----
path <- "Podcasts/"
dir.create(path)

## Get podcasts name ----
podcasts <- list_podcasts()
podcast  <- podcasts[3, ]

## Retrieve episodes information ----
tab <- get_metadata(podcast    = podcast$"podcast", 
                    start_date = podcast$"start_date", 
                    end_date   = podcast$"end_date", 
                    path = path)
}
}
