% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_podcasts.R
\name{list_podcasts}
\alias{list_podcasts}
\title{List of France Inter Podcasts}
\usage{
list_podcasts()
}
\value{
A three-columns data frame with:
\itemize{
\item \code{podcast}: the label of the podcast;
\item \code{start_date}: the date of the first episode.
\item \code{end_date}: the date of the last episode (can be \code{NA}).
}
}
\description{
Lists a selection of France Inter podcasts.
}
\examples{
list_podcasts()
}
