% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_m3u.R
\name{add_m3u}
\alias{add_m3u}
\title{Create a M3U File from a list of URL}
\usage{
add_m3u(data, podcast, path = ".", na_rm = TRUE)
}
\arguments{
\item{data}{a data frame created by \code{get_metadata()}.}

\item{podcast}{a character of length 1. The name of the M3U file.}

\item{path}{the path to save the M3U file (must exist).}

\item{na_rm}{a logical. If \code{TRUE} (default) remove episodes with incomplete
information.}
}
\value{
The content of the M3U file (optional).
}
\description{
Creates a M3U (MPEG version 3.0 URL) file from a list of URLs. This file can
be read by any multimedia reader (e.g. VLC). These audio/video URLs will be
streamed (not downloaded).
For more information: \url{https://en.wikipedia.org/wiki/M3U}
}
\examples{
\dontrun{
## Create a folder to store results ----
path <- "Podcasts/"
dir.create(path)

## Get podcasts name ----
podcasts <- list_podcasts()
podcast  <- podcasts[3, ]

## Retrieve episodes information ----
tab <- get_metadata(podcast    = podcast$"podcast", 
                    start_date = podcast$"start_date", 
                    end_date   = podcast$"end_date", 
                    path = path)

## Create a M3U playlist ----
x <- add_m3u(tab, podcast$"podcast", path = path)

cat(paste0(x, collapse = "\n"))
}
}
