% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franceinter-package.R
\docType{package}
\name{franceinter-package}
\alias{franceinter}
\alias{franceinter-package}
\title{franceinter: Get France Inter Podcasts}
\description{
Gets France Inter metadata (i.e. date, title, and mp3 url) for a
    selected podcasts (e.g. Tanguy Pastureau, Guillaume Meurice, Waly Dia, Sur 
    les Epaules de Darwin, Ca peut pas faire de mal).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ahasverus/franceinter}
  \item Report bugs at \url{https://github.com/ahasverus/franceinter/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID})

}
\keyword{internal}
