% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Retrieve podcast metadata}
\usage{
get_metadata(podcast, path = ".", na_rm = TRUE)
}
\arguments{
\item{podcast}{a character of length 1. The name of the podcast.}

\item{path}{a character of length 1. The folder to save metadata as CSV.}

\item{na_rm}{a logical. If \code{TRUE} (default) remove episodes with incomplete
information.}
}
\value{
A four-columns data frame with:
\itemize{
\item \code{date}: the date of the episode;
\item \code{title}: the title of the episode;
\item \code{duration}: the duration of the episode (in seconds);
\item \code{file_url}: the URL of the mp3.
}
}
\description{
Retrieves podcast metadata (i.e. date, title, mp3 duration and URL).
Information will be saved in a CSV file. Each time this function is called
only new episodes will be added.
}
\examples{
\dontrun{
## Create a folder to store results ----

path <- "Podcasts"
dir.create(path)


## Retrieve episodes information ----

tab <- get_metadata("un-ete-avec-homere", path = "inst")
}
}
